requirements.txt
Django==1.9.4
argparse==1.2.1
dj-database-url==0.4.0
dj-static==0.0.6
django-annoying==0.9.0
django-crispy-forms==1.6.0
django-registration-redux==1.4
django-scheduler==0.7.5
django-timedeltafield==0.7.10
django-toolbelt==0.0.1
future==0.15.2
gunicorn==19.4.5
icalendar==3.9.2
psycopg2==2.6.1
python-dateutil==2.5.0
pytz==2016.1
six==1.10.0
static3==0.6.1
whitenoise==3.0
wsgiref==0.1.2
